prediction_app/requirements-deploy.txt
flask
tensorflow~=2.11.0
numpy~=1.24.1
pillow==9.2.0
gunicorn